% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daacsR-package.R
\docType{package}
\name{daacsR-package}
\alias{daacsR}
\alias{daacsR-package}
\title{daacsR: Diagnostic Assessment and Achievement of College Skills R Package}
\description{
Utility functions to access DAACS database and manipulate data for analysis.
}
\author{
\strong{Maintainer}: Jason Bryer \email{jason.bryer@cuny.edu} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

}
\keyword{internal}
