% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDAACSConnection.R
\name{getDAACSConnection}
\alias{getDAACSConnection}
\title{Returns a connection to a DAACS database.}
\usage{
getDAACSConnection(username, password, host, port, db)
}
\description{
Returns a connection to a DAACS database.
}
