% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getItemResults.R
\name{getItemResults}
\alias{getItemResults}
\title{Returns individal item responses.}
\usage{
getItemResults(
  daacs_conn,
  student_id,
  assessments = c("COLLEGE_SKILLS", "MATHEMATICS", "WRITING", "READING"),
  ...
)
}
\arguments{
\item{daacs_conn}{a connection to the mongolite database.}

\item{student_id}{(optional) can be a character vector to limit the students
data to return. Of omitted, return all student data.}

\item{assessments}{list of assessments to include results from.}
}
\value{
a data.frame where each row represents a student response to an
individual question.
\describe{
\item\link{username} student's username in DAACS.
\item\link{assessment} which assessment this item is from.
\item\link{takenDate} the time the student completed the question.
\item\link{stem} the question stem
\item\link{response} the student's answer to the question as a character (i.e.
strongly agree for SRL, A through D for multiple choice).
\item\link{score} the score for the student.
}
}
\description{
Returns individal item responses.
}
