% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStudentResults.R
\name{getStudentResults}
\alias{getStudentResults}
\title{Returns student results}
\usage{
getStudentResults(
  daacs_conn,
  student_id,
  assessments = c("COLLEGE_SKILLS", "MATHEMATICS", "WRITING", "READING"),
  calc_overall = TRUE,
  domains = FALSE,
  ...
)
}
\arguments{
\item{daacs_conn}{a connection to the mongolite database.}

\item{student_id}{(optional) can be a character vector to limit the students
data to return. Of omitted, return all student data.}

\item{assessments}{list of assessments to include results from.}

\item{calc_overall}{returns the overall numeric score which is the average
of item responses.}

\item{domains}{whether to include all of the domain scores within each
assessment in the returned data.frame.}

\item{...}{other parameters passed to mongolite.}
}
\value{
a data.frame with student results.
\describe{
\item\link{username} username from DAACS
\item\link{firstName} student's first name
\item\link{lastName} student's last name
\item\link{assessmentCategory} the domain (i.e. COLLEGE_SKILLS, MATHEMATICS, etc.)
\item\link{assessmentLabel}
\item\link{takenDate}
\item\link{status}
\item\link{takenDate}
\item\link{overallScore}
\item\link{SUBDOMAIN SCORES}
}
}
\description{
Returns student results
}
\examples{
\dontrun{
getStudentResults() # Returns all students
getStudentResults(c('jbryer@excelsior.edu', 'angela.lui@cuny.edu'))
}
}
